int to long

+++ AdventOfCode2021/Day1.cs
@@ -8,9 +8,9 @@ namespace AdventOfCode2021
 {
     public class Day1
     {
-        public int Main(string part, string path)
+        public long Main(string part, string path)
         {
-            int answer = 0;
+            long answer = 0;
             if (part == "a")
             {
                 answer = Day1a(path);
